+++ zcash_client_sqlite/src/wallet/sapling.rs
@@ -556,7 +556,7 @@ pub(crate) mod tests {
         let to_extsk = ExtendedSpendingKey::master(&[]);
         let to: Address = to_extsk.default_address().1.into();
         let request = zip321::TransactionRequest::new(vec![Payment {
-            recipient_address: to,
+            recipient_address: to.to_zcash_address(&st.network()),
             amount: NonNegativeAmount::const_from_u64(10000),
             memo: None, // this should result in the creation of an empty memo
             label: None,
@@ -1273,7 +1273,7 @@ pub(crate) mod tests {
         let req = TransactionRequest::new(vec![
             // payment to an external recipient
             Payment {
-                recipient_address: Address::Sapling(addr2),
+                recipient_address: Address::Sapling(addr2).to_zcash_address(&st.network()),
                 amount: amount_sent,
                 memo: None,
                 label: None,
@@ -1282,7 +1282,7 @@ pub(crate) mod tests {
             },
             // payment back to the originating wallet, simulating legacy change
             Payment {
-                recipient_address: Address::Sapling(addr),
+                recipient_address: Address::Sapling(addr).to_zcash_address(&st.network()),
                 amount: amount_legacy_change,
                 memo: None,
                 label: None,
@@ -1396,7 +1396,8 @@ pub(crate) mod tests {
 
         // This first request will fail due to insufficient non-dust funds
         let req = TransactionRequest::new(vec![Payment {
-            recipient_address: Address::Sapling(dfvk.default_address().1),
+            recipient_address: Address::Sapling(dfvk.default_address().1)
+                .to_zcash_address(&st.network()),
             amount: NonNegativeAmount::const_from_u64(50000),
             memo: None,
             label: None,
@@ -1421,7 +1422,8 @@ pub(crate) mod tests {
         // This request will succeed, spending a single dust input to pay the 10000
         // ZAT fee in addition to the 41000 ZAT output to the recipient
         let req = TransactionRequest::new(vec![Payment {
-            recipient_address: Address::Sapling(dfvk.default_address().1),
+            recipient_address: Address::Sapling(dfvk.default_address().1)
+                .to_zcash_address(&st.network()),
             amount: NonNegativeAmount::const_from_u64(41000),
             memo: None,
             label: None,